[MCUX-87815][MCXN] fix device driver certc issues
1. fix device driver certc issues

+++ MCXN/MCXN947/drivers/fsl_clock.h
@@ -545,7 +545,7 @@ typedef enum _clock_name
 
 #define GET_ID_ITEM(connection)      ((connection)&0xFFFFU)
 #define GET_ID_NEXT_ITEM(connection) ((connection) >> 16U)
-#define GET_ID_ITEM_MUX(connection)  (((uint16_t)connection) & 0xFFFU)
+#define GET_ID_ITEM_MUX(connection)  ((uint32_t)(connection) & 0xFFFU)
 #define GET_ID_ITEM_SEL(connection)  ((uint8_t)(((((uint32_t)(connection)&0xF000U) >> 12U) - 1U) & 0xFFU))
 #define GET_ID_SELECTOR(connection)  ((connection)&0xF000000U)
